Allow entering partial servings
This change is necessary because you don't always consume whole
servings.

This change changes the servings eaten variables from an int to a double
to allow partial servings.

There are no side effets to this change.

No link to an issue is available.

+++ NutritionalTracker/Commands/AddRecipeToDiaryCommand.cs
@@ -5,7 +5,7 @@ namespace NutritionalTracker.Commands
     public sealed class AddRecipeToDiaryCommand : ICommand
     {
         public int RecipeId { get; set; }
-        public int NumberOfServingsConsumed { get; set; }
+        public double NumberOfServingsConsumed { get; set; }
         public byte MealId { get; set; }
         public DateTime ConsumedDate { get; set; }
     }
